feat: create index page that redirects to page with user language
